requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cssselect==1.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fake-useragent==1.2.1
fastapi==0.103.1
feedparser==6.0.10
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
orjson==3.9.7
packaging==23.1
pandas==2.1.0
parse==1.19.1
pluggy==1.3.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.2
watchfiles==0.20.0
websockets==10.4
yarl==1.9.2
zipp==3.16.2
